stocks/requirements.txt
requests
finnhub-python
pandas
numpy
matplotlib
yfinance